Extract UnicodePlot::Utils module
