+++ lib/unicode_plot.rb
@@ -1,5 +1,6 @@
 require 'unicode_plot/version'
 
+require 'unicode_plot/utils'
 require 'unicode_plot/styled_printer'
 require 'unicode_plot/value_transformer'
 require 'unicode_plot/renderer'
